Handle valid formatted but missing PU numbers
This also tidies up the code for trying multiple variations of the pun.
Previously we were using a while loop and a regular expression to try
the different variations of the pu number. I've changed things to
pre-compute the variations of the pu number and then try those in turn,
returning for the first successful one.